+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/TagsExtractingProcessor.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -23,9 +23,11 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
 import com.walmartlabs.concord.project.InternalConstants;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
+import com.walmartlabs.concord.server.process.ProcessKey;
 
 import javax.inject.Named;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Named
 public class TagsExtractingProcessor implements PayloadProcessor {
@@ -40,29 +42,38 @@ public class TagsExtractingProcessor implements PayloadProcessor {
             return chain.process(payload);
         }
 
-        Set<String> tags;
-
-        if (v instanceof String[]) {
-            String[] as = (String[]) v;
-            tags = new HashSet<>(as.length);
-            Collections.addAll(tags, as);
-        } else if (v instanceof Collection) {
-            Collection c = (Collection) v;
-            tags = new HashSet<>(c.size());
-
-            for (Object o : c) {
-                if (o instanceof String) {
-                    tags.add((String) o);
-                } else {
-                    throw new ProcessException(payload.getProcessKey(), "Process tag must be a string value: " + o);
-                }
-            }
-        } else {
-            throw new ProcessException(payload.getProcessKey(), "Process tags must be an array of string values: " + v);
-        }
-
+        Set<String> tags = parse(payload.getProcessKey(), v);
         payload = payload.putHeader(Payload.PROCESS_TAGS, tags);
 
         return chain.process(payload);
+    }
+
+    private static Set<String> parse(ProcessKey pk, Object v) {
+        if (v instanceof String) {
+            String[] as = ((String) v).split(",");
+            return trim(as);
+        } else if (v instanceof String[]) {
+            String[] as = (String[]) v;
+            return trim(as);
+        } else if (v instanceof Collection) {
+            Set<String> result = new HashSet<>();
+
+            Collection c = (Collection) v;
+            for (Object o : c) {
+                if (!(o instanceof String)) {
+                    throw new ProcessException(pk, "Process tag must be a string value: " + o);
+                }
+
+                result.add((String) o);
+            }
+
+            return result.stream().map(String::trim).collect(Collectors.toSet());
+        } else {
+            throw new ProcessException(pk, "Process tags must be an array of string values: " + v);
+        }
+    }
+
+    private static Set<String> trim(String[] as) {
+        return Arrays.stream(as).map(String::trim).collect(Collectors.toSet());
     }
 }