+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Changed
 
+- concord-console: add process tags to the process status page;
+- concord-server: process `tags` can now be specified using a comma-separated
+startup argument, e.g. `curl -F tags=x,y,z`;
 - concord-server: store repository info (commit ID, commit message,
 etc) early in the pipeline to preserve the data in case or process
 startup errors (e.g. bad syntax).